Split `test_nurse_consent.py`
This splits up the file in to two separate files,
`test_parent_consent.py` and `test_self_consent.py`. This is because
we'll be adding more self-consent tests and that file will get quite
long.

Jira-Issue: MAV-5891
